+++ Whisky/Views/ContentView.swift
@@ -128,30 +128,6 @@ struct ContentView: View {
             if !WhiskyWineInstaller.isWhiskyWineInstalled() {
                 showSetup = true
             }
-            let task = Task.detached {
-                await WhiskyWineInstaller.shouldUpdateWhiskyWine()
-            }
-            let updateInfo = await task.value
-            if updateInfo.0 {
-                let alert = NSAlert()
-                alert.messageText = String(localized: "update.whiskywine.title")
-                alert.informativeText = String(
-                    format: String(localized: "update.whiskywine.description"),
-                    String(WhiskyWineInstaller.whiskyWineVersion()
-                        ?? SemanticVersion(0, 0, 0)),
-                    String(updateInfo.1)
-                )
-                alert.alertStyle = .warning
-                alert.addButton(withTitle: String(localized: "update.whiskywine.update"))
-                alert.addButton(withTitle: String(localized: "button.removeAlert.cancel"))
-
-                let response = alert.runModal()
-
-                if response == .alertFirstButtonReturn {
-                    WhiskyWineInstaller.uninstall()
-                    showSetup = true
-                }
-            }
         }
     }